Update Build Version handling
Prior to this commit, the build process grabbed the latest version from
the git_remote and built using the production branch.  *IF* there were
updates between the time you tagged the release you wanted and this
build kicking off you could wind up with newer commits.  These updates
cause the git version in the local SCM to be reset back to the build
version requested.