Update role for this all-in-one SETeam master
The role name in the control-repo for this server has changed, to
accomodate an additional role. The old name became ambiguous and so a
new name was selected for clarity. The old name continued to work
temporarily.

This commit updates the tse-master-builder code to use the new role
name, so that the old role name can be removed from the control-repo.